uri에서 오버플로우 및 언더플로우가 날 가능성이 있는 타입 변경, 일부 API에서 size가 필수 파라미터였던 문제 수정
(cherry picked from commit bb96ac47a903dc0487c3f073ff1092bcbb507bc8)

+++ src/main/java/kr/milibrary/mapper/BookMapper.java
@@ -31,5 +31,5 @@ public interface BookMapper {
     List<Book> searchBooks(@Param("criteria") Book.SearchCriteria criteria);
 
     @Select("SELECT COUNT(*) FROM milibrary.books;")
-    int getTotalCount();
+    long getTotalCount();
 }
